fix(desktop): restore remote artifact rendering
